day18/hirst-painting/requirements.txt
colorgram.py==1.2.0
Pillow==8.1.2
